Update pluck for migrating existing data
This updates the "pluck" implementation to fall back to the original,
unencrypted columns during the time of migrating existing data.

While attributes are set to "migrating: true", the newly introduced
"_ciphertext" columns of old records may still be empty until migrating
the data from the original unencrypted columns is finished. Therefore
"pluck" keeps on using the original unencrypted columns until
"migrating: true" is removed from the attribute definition and we can be
sure that the encrypted columns are the source of truth.

closes #79